+++ doc-site/pages/examples/viewmodel/fields/JsonField/form-usage.tsx
@@ -4,18 +4,15 @@
  * This example shows the default widget that will be used in a [Form](doc:Form)
  * when using [@prestojs/ui-antd](/docs/ui-antd). See [getWidgetForField](doc:getWidgetForField).
  *
- * The default widget is [EmailWidget](doc:CharWidget)
+ * The default widget is [JsonWidget](doc:JsonWidget). You can pass `format` and `parse` to the form field in
+ * order to handle transforming value back and forth from a JSON string to it's parsed value.
  *
- * If `maxLength` is specified the widget will limit the length of entered text.
- *
- * Any extra widget props can be defined at the field level in the `widgetProps` option (eg.
- * `placeholder` in this example).
- *
- * @wide
+ * Any extra widget props can be defined at the field level in the `widgetProps` option (e.g. `rows` in the example
+ * below).
  */
 import { Form } from '@prestojs/final-form';
 import { AntdUiProvider, FormItemWrapper, FormWrapper, getWidgetForField } from '@prestojs/ui-antd';
-import { EmailField, IntegerField, viewModelFactory } from '@prestojs/viewmodel';
+import { IntegerField, JsonField, viewModelFactory } from '@prestojs/viewmodel';
 import { Button } from 'antd';
 import 'antd/dist/antd.min.css';
 import React from 'react';
@@ -23,17 +20,16 @@ import React from 'react';
 class ExampleModel extends viewModelFactory(
     {
         id: new IntegerField(),
-        email: new EmailField({
+        data: new JsonField({
             helpText: 'Please enter your email address',
-            widgetProps: {
-                placeholder: 'Enter your email address',
-            },
+            widgetProps: { rows: 5 },
         }),
     },
     { pkFieldName: 'id' }
 ) {}
 
 export default function FormUsage() {
+    const record = new ExampleModel({ id: 1, data: '{"name": "Gandalf"}' });
     return (
         <React.Suspense fallback="Loading...">
             <AntdUiProvider
@@ -42,8 +38,11 @@ export default function FormUsage() {
                 formComponent={FormWrapper}
             >
                 <div className="grid grid-cols-1 gap-4 w-full">
-                    <Form onSubmit={data => console.log(data)}>
-                        <Form.Item field={ExampleModel.fields.email} />
+                    <Form
+                        onSubmit={({ data }) => console.log(ExampleModel.fields.data.parse(data))}
+                        initialValues={{ data: ExampleModel.fields.data.format(record.data) }}
+                    >
+                        <Form.Item field={ExampleModel.fields.data} />
                         <Form.Item wrapperCol={{ offset: 6 }}>
                             <Button type="primary" htmlType="submit">
                                 Submit (check console)