+++ js-packages/@prestojs/ui-antd/src/widgets/JsonWidget.tsx
@@ -1,25 +1,43 @@
 import { WidgetProps } from '@prestojs/ui';
 import { Input } from 'antd';
 import { TextAreaRef } from 'antd/lib/input/TextArea';
-import React from 'react';
+import React, { ComponentProps } from 'react';
 
 /**
  * @expand-properties
- * @hide-properties choices asyncChoices
+ * @hide-properties meta
  */
-type JsonWidgetProps = WidgetProps<string, HTMLTextAreaElement>;
+type JsonWidgetProps = Omit<
+    WidgetProps<string | null, HTMLTextAreaElement>,
+    'choices' | 'asyncChoices'
+> &
+    Omit<ComponentProps<typeof Input.TextArea>, 'onChange' | 'value'>;
+
 /**
- * See [Input.TextArea](https://ant.design/components/input/#Input.TextArea) for props available
+ * Form widget for JSON values that renders as a [Input.TextArea](https://4x.ant.design/components/input/#Input.TextArea).
  *
- * Renders as a TextArea
+ * This is the [default widget](doc:getWidgetForField) used for [JsonField](doc:JsonField)
+ *
+ * <Usage type="widget" widgetName="JsonWidget">
+ * ```js
+ * function JsonWidgetExample() {
+ * const [value, setValue] = useState("");
+ *  return <JsonWidget input={{ onChange({ target: { value } }) {
+ *  setValue(value)
+ * }, value}} />
+ * }
+ * ```
+ * </Usage>
  *
  * @extract-docs
  * @menu-group Widgets
  * @forward-ref
+ * @hide-properties meta
  */
 function JsonWidget(props: JsonWidgetProps, ref): React.ReactElement {
     const { input, meta, ...rest } = props;
-    return <Input.TextArea ref={ref} {...input} {...rest} />;
+    const { value, ...restInput } = input;
+    return <Input.TextArea ref={ref} value={value ?? ''} {...restInput} {...rest} />;
 }
 
 export default React.forwardRef<TextAreaRef, WidgetProps<string, HTMLTextAreaElement>>(JsonWidget);