+++ js-packages/@prestojs/ui-antd/src/widgets/NumberWidget.tsx
@@ -1,26 +1,49 @@
 import { WidgetProps } from '@prestojs/ui';
 import { InputNumber } from 'antd';
-import React from 'react';
+import React, { ComponentProps } from 'react';
+
+type ValueType = string | number;
 
 /**
  * @expand-properties
- * @hide-properties choices asyncChoices
+ * @hide-properties meta
  */
-type NumberWidgetProps = WidgetProps<number, HTMLInputElement>;
+export type NumberWidgetProps<T extends ValueType = ValueType> = Omit<
+    WidgetProps<T | null, HTMLInputElement>,
+    'choices' | 'asyncChoices'
+> &
+    Omit<ComponentProps<typeof InputNumber>, 'onChange' | 'value'>;
 
 /**
- * See [InputNumber](https://ant.design/components/input-number/) for props available
+ * Form widget for string values that renders as a [InputNumber](https://4x.ant.design/components/input-number/).
+ *
+ * This is the [default widget](doc:getWidgetForField) used for [NumberField](doc:NumberField)
+ *
+ * <Usage type="widget" widgetName="NumberWidget">
+ * ```js
+ * function NumberWidgetExample() {
+ * const [value, setValue] = useState(null);
+ *  return <NumberWidget input={{ onChange(nextValue) {
+ *  setValue(nextValue)
+ * }, value}} />
+ * }
+ * ```
+ * </Usage>
  *
  * @extract-docs
  * @menu-group Widgets
  * @forward-ref
+ * @hide-properties meta
  */
-function NumberWidget(
-    props: NumberWidgetProps,
+function NumberWidget<T extends ValueType = number>(
+    props: NumberWidgetProps<T>,
     ref: React.RefObject<HTMLInputElement>
 ): React.ReactElement {
     const { input, meta, ...rest } = props;
     return <InputNumber ref={ref} {...input} {...rest} />;
 }
 
-export default React.forwardRef(NumberWidget);
+export default React.forwardRef(NumberWidget) as <T extends ValueType = number>(
+    props: NumberWidgetProps<T>,
+    ref: React.RefObject<HTMLInputElement>
+) => React.ReactElement;